requirements.txt
click>=8.1.7
requests>=2.31.0
aiohttp>=3.8.5
rich>=13.3.0
tqdm>=4.66.1
PyYAML>=6.0
python-dotenv>=1.0.0
fastmcp>=2.13.2